refactor: simplify plugin install, correct a-la-carte exports
