+++ packages/vuetify/src/entry-lib.ts
@@ -1,4 +1,5 @@
 export * from './components'
 export * from './directives'
 export { default as colors } from './util/colors'
-export { default } from './framework'
+export { install as default } from './install'
+export { useVuetify } from './framework'
